Pull out libraries and tests
